Kafka_Lab/requirements.txt
confluent-kafka==2.8.0
matplotlib==3.7.5
numpy==1.24.4
scikit-learn==1.3.2
pandas
pyyaml
plotly
streamlit